DinucleotideForces

Documentation for DinucleotideForces, a Julia package to computed dinucleotide forces in nucleic-acid sequences.

Installation

This package is not registered. Install with:

import Pkg
Pkg.add(url="https://github.com/adigioacchino/DinucleotideForces.jl")

Dinucleotide forces: definition

Nucleotide/dinucleotide forces for a sequence are parameters that capture pressures to increase or decrease the usage of a given nucleotide/dinucleotide in the sequence. So a positive force on a nucleotide corresponds to a pressure to have more occurrences of the given dinucleotide, while a negative pressure corresponds to a lack of occurrences of the dinucleotide in the considered sequence. Forces are inferred through a maximum entropy model. For a more formal definition give a look at this paper.

How to cite this package

If this package is used for research purposes, please consider citing this and this papers.

Examples

Basic usage

We will take sub-sequence from the Influenza H5N1 PB2 segment (strain used: A/Anhui/1/2005) and we will use it as working example to test the function of the scripts. We start by loading the script and the sequence:

using DinucleotideForces
example_seq = "ATGGAGAGAATAAAAGAATTAAGGGATCTAATGTCACAGTCCCGCACTCGCGAGATACTAACAAAAACCACTGTGGACCATATGGCCATAATCAAGAAGTACACATCAGGAAGACAAGAGAAGAACCCTGCTCTCAGAATGAAATGGATGATGGCAATGAAATATCCAATCACAGCGGACAAGAGAATAACAGAGATGATTCCTGAAAGGAATGAACAAGGGCAGACGCTCTGGAGCAAGACAAATGATGCCGGATCGGACAGGTTGATGGTGTCTCCCTTAGCTGTAACTTGGTGGAATAGGAATGGGCCGACGACAAGTGCAGTCCATTATCCAAAGGTTTACAAAACATACTTTGAGAAGGCT"

Let's start by computing the force on the CpG motif along the full genome: this can be done with the only exported function of the package, DimerForce, as shown in the following code block.

julia> motifs = ["CG"];
julia> DimerForce(example_seq, motifs)Dict{String, Float64} with 5 entries:
  "A"  => -1.06703
  "T"  => -1.73119
  "C"  => -1.547
  "CG" => -0.944914
  "G"  => -1.32423

The resulting output will be a dictionary where the force is associated to each motif specified (in this case only the "CG" motif).

When no nucleotide biases (frequencies) are specified, as above, the corresponding forces are inferred together with the forces. We will use the word fields to indicate forces on nucleotides, while forces will be from now on restricted to dinucleotides. Notice that $\sum_n e^{h_n} = 1$, where $n \in \{A,C,G,T\}$. This is the result of a gauge choice, because in general the model probabilities are invariant under the transformation $ hn \to hn+K$, which can then be used to make the fields interpretable as the logarithm of a frequency (more on this is discussed later in section Focus on gauge choices).

A user-specified bias can be given, and in this case fields are not inferred:

julia> nt_bias = [0.25, 0.25, 0.25, 0.25]; # probs for A, C, G, T
julia> motifs = ["CG"];
julia> DimerForce(example_seq, motifs; freqs=nt_bias)Dict{String, Float64} with 1 entry:
  "CG" => -1.06519

The package easily allows to compute forces on two or more dinucleotides:

julia> motifs = ["CG", "TA"];
julia> DimerForce(example_seq, motifs)Dict{String, Float64} with 6 entries:
  "A"  => -1.0215
  "T"  => -1.58094
  "C"  => -1.64745
  "CG" => -0.864869
  "G"  => -1.42037
  "TA" => -0.754889

julia> nts = ["A", "C", "G", "T"];
julia> motifs = [a*b for a in nts for b in nts];
julia> DimerForce(example_seq, motifs)Dict{String, Float64} with 20 entries:
  "T"  => -1.76995
  "C"  => -1.53967
  "CC" => 0.0
  "GC" => 0.203128
  "GG" => 0.0
  "CG" => -1.25295
  "AT" => 0.0
  "A"  => -1.04555
  "CA" => 0.206446
  "TG" => 0.0859823
  "TA" => -0.144103
  "GT" => 0.0
  "G"  => -1.33292
  "GA" => 0.790941
  "TT" => 0.0
  "AC" => -0.185393
  "CT" => 0.0
  "AA" => 0.0
  "AG" => -0.639832
  "TC" => 0.403223

Look at how the CpG force changed in the cases: this is due to the fact that different motifs interact.

Focus on gauge choices

When the full set of fields and forces is inferred, the system of equations solved to obtain these parameters is underdetermined. This means that a gauge choice must be made, for instance by setting some forces to zero by passing less than 16 dinucleotides as motifs. If such a choise is not made, the package selects the gauge so that:

the exponential of the fields sum to 1;

the forces for dinucleotides of the form NN and those of the form NT are put to zero.

Notice that the script allows for flexible choices: if not all the dinucleotides are given, the gauge is always chosen such that the maximum possible number of the nucleotides of the form NN and NT have forces equal to zero, as one can check by running the next code blocks.

julia> motifs = ["AC", "AG", "AT", "CA", "GA", "TA"];
julia> DimerForce(example_seq, motifs)Dict{String, Float64} with 10 entries:
  "AC" => -0.412762
  "AT" => 0.0
  "A"  => -0.999029
  "T"  => -1.53904
  "C"  => -1.67771
  "AG" => -0.373976
  "CA" => 0.563492
  "G"  => -1.46804
  "TA" => -0.365219
  "GA" => 0.652723

julia> motifs = [
           "AC","AG","AT","CA",
           "CG","CT","GA","GC",
           "GT","TA","TC","TG",
           ];
julia> DimerForce(example_seq, motifs)Dict{String, Float64} with 16 entries:
  "T"  => -1.84946
  "C"  => -1.53652
  "GC" => 0.211443
  "CG" => -1.23487
  "AT" => 0.0
  "A"  => -1.0114
  "CA" => 0.175394
  "TG" => 0.211296
  "G"  => -1.33241
  "TA" => -0.0679201
  "GT" => 0.0
  "GA" => 0.75677
  "AC" => -0.178653
  "CT" => 0.0
  "AG" => -0.626449
  "TC" => 0.521892

Notice that when dinucleotides are not given, this is equivalent to fix to 0 their forces. Depending on which dinucleotides are not given, this might or not be a specific gauge - in other words, it might or not result in an equivalent model. For instance, the following three cells result in equivalent models (although the third has different parameters because it is in another gauge), while the fourth does not:

julia> motifs = ["AC", "AG", "AT", "CA", "GA", "TA"];
julia> DimerForce(example_seq, motifs)Dict{String, Float64} with 10 entries:
  "AC" => -0.412762
  "AT" => 0.0
  "A"  => -0.999029
  "T"  => -1.53904
  "C"  => -1.67771
  "AG" => -0.373976
  "CA" => 0.563492
  "G"  => -1.46804
  "TA" => -0.365219
  "GA" => 0.652723

julia> motifs = ["AC", "AG", "CA", "GA", "TA"];
julia> DimerForce(example_seq, motifs)Dict{String, Float64} with 9 entries:
  "AC" => -0.412762
  "A"  => -0.999029
  "T"  => -1.53904
  "C"  => -1.67771
  "AG" => -0.373976
  "CA" => 0.563492
  "G"  => -1.46804
  "TA" => -0.365219
  "GA" => 0.652723

julia> motifs = ["AG", "AT", "CA", "GA", "TA"];
julia> DimerForce(example_seq, motifs)Dict{String, Float64} with 9 entries:
  "AT" => 0.537838
  "A"  => -0.998373
  "T"  => -1.56506
  "C"  => -1.65424
  "AG" => 0.097891
  "CA" => 0.0930896
  "G"  => -1.46444
  "TA" => -0.835621
  "GA" => 0.182321

julia> motifs = ["AG", "CA", "GA", "TA"];
julia> DimerForce(example_seq, motifs)Dict{String, Float64} with 8 entries:
  "A"  => -1.00421
  "T"  => -1.46014
  "C"  => -1.76025
  "AG" => -0.174972
  "CA" => 0.365081
  "G"  => -1.47203
  "TA" => -0.563632
  "GA" => 0.454311

In some cases (notably when all the dinucleotides are given), fixing fields is equivalent to fix (part of) a gauge, so the model is independendent on this choice (although the values of the inferred parameters depend on that).

Inferring forces from multiple sequences

The package is able to infer fields and forces from a set of sequences:

julia> motifs = ["CG"];
julia> seqs = [example_seq, example_seq];
julia> DimerForce(seqs, motifs)Dict{String, Float64} with 5 entries:
  "A"  => -1.06703
  "T"  => -1.73119
  "C"  => -1.547
  "CG" => -0.944914
  "G"  => -1.32423

The sequences do not need to have the same length. However, the values of the inferred parameters weakly depend on the sequence length, that has to be provided

julia> motifs = ["CG"];
julia> seqs = [example_seq, example_seq[1:100]];
julia> DimerForce(seqs, motifs, 2500)Dict{String, Float64} with 5 entries:
  "A"  => -1.01189
  "T"  => -1.71435
  "C"  => -1.54984
  "CG" => -0.725426
  "G"  => -1.41015

or it will arbitarily set by the package to the standard value of 5000

julia> motifs = ["CG"];
julia> seqs = [example_seq, example_seq[1:100]];
julia> DimerForce(seqs, motifs)Warning: Lmodel is not specified and the sequence set provided contains sequences of different lengths. Using an arbitrary value of L=5000.
Dict{String, Float64} with 5 entries:
  "A"  => -1.01229
  "T"  => -1.71423
  "C"  => -1.54956
  "CG" => -0.726089
  "G"  => -1.40989

Notice, in the last two examples, how the changes due to the different lengths used during the inference of the parameters only marginally affect their values.

Computing energies and loglikelihoods

Given a sequence, the package allows to compute easily its energy assigned by a model

julia> model = DimerForce(example_seq, ["CG"]);
julia> DinucleotideForces.ComputeEnergy(example_seq, model)-502.5299370908549

and its log-likelihood

julia> DinucleotideForces.ComputeLoglikelihood(example_seq, model)-489.14789197001

To make sense of this large negative log-likelihood, consider that in the uniform case we have $4^L$ sequences, so the log-likelihood of each of them would be:

julia> - length(example_seq) * log(4)-507.38373616987997

Notice that the function ComputeLoglikelihood computes the partition function each time it is runned. This can slow down the computation when the function is called many times with the same model and to prevent this the log of the partition function can be passed directly:

julia> model = DimerForce(example_seq, ["CG", "AC"]);
julia> ks = keys(model);
julia> k1 = [k for k in ks if length(k)==1];
julia> k2 = [k for k in ks if length(k)==2];
julia> fields = [model[k] for k in k1];
julia> forces = [model[k] for k in k2];
julia> logZ = DinucleotideForces.eval_log_Z(fields, forces, k2, length(example_seq));ERROR: UndefVarError: eval_log_Z not defined
julia> DinucleotideForces.ComputeLoglikelihood(example_seq, model; logZ=logZ)ERROR: UndefVarError: logZ not defined

Sampling sequences

The framework used by this package introduced a probability distribution over the sequences given a set of fields and/or forces. Sequences can be sampled from this distribution exactly (that is, without using any Markov-chain based algorithm such as Metropolis) using the simple topology of the graph of interactions between nucleotides (a tree) and the fact that the partition function can be computed exactly. The funcion SampleSequence of the package implements this idea, and it is very easy to use, its only arguments being the model that has to be used to sample from and the length of the desired sequence:

julia> nts = ["A", "C", "G", "T"];
julia> motifs = [a*b for a in nts for b in nts];
julia> model = DimerForce(example_seq, motifs);
julia> DinucleotideForces.SampleSequence(model, 1000)"GACGATTAGTTCCATTACTGAAGGATAAATGAACAAATAAGAGAGAGGCACCAGACGCCGATACTCCTGTGAGGGGCTACCGAAAACCACACGCAAACCAGTACACGAGCTTTGCCATAAAGCTGGACGAATCCCAAGATAAACAACAGGGTGCCATGGGGACAGAAATCAAGAAACTTAGTGAGCGGACCAAAGCTCAAAGGCACAAATCAACTAGTGGGAGAGTAAAGGGCCCTTTAAAATTAAGAATGTTCACCGCGGCCAGATGAAA" ⋯ 459 bytes ⋯ "CCTCCAAACCCATGCCCCAGCCATGAGGGGGCCCCATAGGAAAACGGGGTCACACCCTGAATAAATTCAGGTCAGATAGAAAAATTACAGAGAGGGGCGGAGATTAACGGAGAGGACTCAAGTGACGAGATGCATAATGGATGACGCTGAGGGACATCTGCACGCATATGTGCTTCTGGAAAATTCCCCCAAGAAGATTGCAGAGTTAGGATCAACTCTTGCCAATTCCACTACGGGGGAGGATTTGTACCCACCAGGTGTGAGGGAGGC"

Known issues

There are issues with the add_pseudocount argument of DimerForce. In principle, using add_pseudocount=true should add a single pseudocount to the number of each motif inferred (nucleotide or di-nucleotide). However, from the output of the following two examples it looks that something is not working properly

julia> nts = ["A", "C", "G", "T"];
julia> motifs = [a*b for a in nts for b in nts];
julia> DimerForce(example_seq, motifs)Dict{String, Float64} with 20 entries:
  "T"  => -1.76995
  "C"  => -1.53967
  "CC" => 0.0
  "GC" => 0.203128
  "GG" => 0.0
  "CG" => -1.25295
  "AT" => 0.0
  "A"  => -1.04555
  "CA" => 0.206446
  "TG" => 0.0859823
  "TA" => -0.144103
  "GT" => 0.0
  "G"  => -1.33292
  "GA" => 0.790941
  "TT" => 0.0
  "AC" => -0.185393
  "CT" => 0.0
  "AA" => 0.0
  "AG" => -0.639832
  "TC" => 0.403223
julia> DimerForce(example_seq, motifs, add_pseudocount=true)Dict{String, Float64} with 20 entries:
  "T"  => NaN
  "C"  => NaN
  "CC" => NaN
  "GC" => NaN
  "GG" => NaN
  "CG" => NaN
  "AT" => NaN
  "A"  => NaN
  "CA" => NaN
  "TG" => NaN
  "TA" => NaN
  "GT" => NaN
  "G"  => NaN
  "GA" => NaN
  "TT" => 0.0
  "AC" => NaN
  "CT" => NaN
  "AA" => NaN
  "AG" => NaN
  "TC" => NaN

This is even worse if a shorter sequence is used

julia> ex2 = example_seq[1:100]"ATGGAGAGAATAAAAGAATTAAGGGATCTAATGTCACAGTCCCGCACTCGCGAGATACTAACAAAAACCACTGTGGACCATATGGCCATAATCAAGAAGT"
julia> DimerForce(ex2, motifs)Dict{String, Float64} with 20 entries:
  "T"  => -1.88382
  "C"  => -1.41788
  "CC" => 0.0
  "GC" => -0.513471
  "GG" => 0.0
  "CG" => -0.508183
  "AT" => 0.0
  "A"  => -0.980651
  "CA" => 0.147666
  "TG" => 0.250867
  "TA" => 0.262363
  "GT" => 0.0
  "G"  => -1.46664
  "GA" => 0.262805
  "TT" => 0.0
  "AC" => -0.439822
  "CT" => 0.0
  "AA" => 0.0
  "AG" => -0.303644
  "TC" => 0.248225
julia> DimerForce(ex2, motifs, add_pseudocount=true)ERROR: LinearAlgebra.SingularException(1)

API

DinucleotideForces.ComputeEnergy

DinucleotideForces.ComputeLoglikelihood

DinucleotideForces.ComputeNumObservations

DinucleotideForces.DimerForce

DinucleotideForces.DimerForceOnlyForces

DinucleotideForces.DimerForceWithFields

DinucleotideForces.EvaluateLogZ

DinucleotideForces.FixFinalGauge!

DinucleotideForces.GaugeAwayVariables

DinucleotideForces.GenerateMotifMatrix

DinucleotideForces.GenerateNucleotideMatrix

DinucleotideForces.Marginal1GivenPrevious

DinucleotideForces.Marginal1Point

DinucleotideForces.Marginal2Points

DinucleotideForces.PreprocessSequences

DinucleotideForces.SampleSequence

DinucleotideForces.ComputeEnergy — Method

ComputeEnergy(seq::AbstractString, fields_forces::Dict{String, Float64})

Given a sequence seq, the single-nucleotide fields and the dinucleotide forces (as a single dict), compute the energy of this sequence.

DinucleotideForces.ComputeLoglikelihood — Method

ComputeLoglikelihood(seq::AbstractString, fields_forces::Dict{String, Float64}; logZ=missing)

Given a sequence seq, the single-nucleotide fields and the dinucleotide forces (as a single dict), compute the log-likelihood (energy minus log of Z) of this sequence. logZ can be passed directly if pre-computed, otherwise is it computed each time this function is called.

DinucleotideForces.ComputeNumObservations — Method

ComputeNumObservations(seqs::Vector{<:AbstractString}, independent_motifs::Vector{<:AbstractString}, L::Int)

For each motif in independent_motifs, compute the number of observed motifs in each sequence in seqs, then divide by the sequence length, take the average of these intensive fractions over the sequences, and multiply by the model length L.

DinucleotideForces.DimerForce — Function

DimerForce(seq::Union{AbstractString,Vector{<:AbstractString}}, motifs::Vector{<:AbstractString}; L::Union{Int,Missing}=missing, 
                freqs::Union{Vector{Float64},Missing}=missing, tolerance::Float64=0.01, max_iter::Int=100, 
                add_pseudocount=false)

If frequencies are given, return the forces on the motifs 'motifs' computed for sequence 'seq' with the frequency bias given. If frequencies are not given, the local fields (NOT frequencies!) are inferred and returned (in a gauge such that the exp of them sum to 1, so the fields can be interpreted as log of frequencies). Notice that seq is used only to compute the number of observed motifs (and nucleotides if the frequencies are not provided by user). Also notice that if motifs are dinucleotides, only up to 12 of them are independent. Finally, tolerance and maxiter are parameters for the Newton-Raphson algorithm used to solve the system of equations. If addpseudocount, a single pseudocount is added for each observed number of nucleotides and dinucleotides. seq can be a single sequence or a vector of sequences. L can be used to define the length of the modeled sequence. If not given, the length of the sequence will be used if a single sequence or a set of sequences of the same lengths are provided. Otherwise, the default value of 5000 will be used.

DinucleotideForces.DimerForceOnlyForces — Function

DimerForceOnlyForces(seqs::Vector{<:AbstractString}, motifs::Vector{<:AbstractString}, fields::Vector{Float64},
                            Lmodel::Union{Int,Missing}=missing; tolerance::Float64=0.01, max_iter::Int=100, add_pseudocount=true)

DinucleotideForces.DimerForceWithFields — Function

DimerForceWithFields(seqs::Vector{<:AbstractString}, nucleotides::Vector{Char}, motifs::Vector{<:AbstractString}, 
                 Lmodel::Union{Int,Missing}=missing; tolerance::Float64=0.01, max_iter::Int=100, add_pseudocount=true)

DinucleotideForces.EvaluateLogZ — Method

Compute the transfer matrix through GenerateMotifMatrix, generate_freqsM and the forces given, then takes the correct power to obtian Z.

DinucleotideForces.FixFinalGauge! — Method

Returns a new dictionary with all the fields/couplings inferred, such that the exp of the fields sum to 1 and as much information as possible is included in the field, by putting other motifs to 0 (if possible, those of the form AA, CC, GG, TT and those ending with T).

DinucleotideForces.GaugeAwayVariables — Method

This function returns the nucleotides or motifs that have to be inferred after as many variables as possible are fixed to 0 through gauge transformations.

DinucleotideForces.GenerateMotifMatrix — Method

Return the matrices used for the computation of the partition function for each motif in motifs. In particular, M_{ij} is the number of times the given motif appears in n(i) + n(j), with n(i) the i-th nucleotide.

DinucleotideForces.GenerateNucleotideMatrix — Method

Return a matrix with the informations about the nt frequences, so that when it is multiplied elementwise with that obtained through GenerateMotifMatrix (after implementing the exponentiation with the force, see EvaluateLogZ), the transfer matrices used to compute Z are finally obtained.

DinucleotideForces.Marginal1GivenPrevious — Method

Marginal1GivenPrevious(Marginal2Points::Dict{String, Float64}, fixed::AbstractString; first::Bool=true)

Given the 2point marginal, fix the first symbol (if first=true, otherwise fix the second) to obtain the probability of the second given that the first is equal to fixed (or viceversa if first=false).

DinucleotideForces.Marginal1Point — Method

Marginal1Point(Marginal2Points::Dict{String, Float64}; first::Bool=true)

Given the 2point marginal, sum over the second symbol (if first=true, otherwise over the first) to obtain the 1-point marginal.

DinucleotideForces.Marginal2Points — Method

Marginal2Points(fields_forces::Dict{String, Float64}, L::Int, pos::Int)

Given a model specified by fields and forces and sequence length, compute the 2-point marginal in position (i-1, i). The result is returned as a dictionary "s(i-1) si" => probability.

DinucleotideForces.PreprocessSequences — Method

PreprocessSequences(seqs_in::Vector{<:AbstractString})

Pre-process the vector of sequences seqs_in, so that sequences only contains letters A, C, G, T. The pre-processed sequence vector is the returned.

DinucleotideForces.SampleSequence — Function

SampleSequence(fields_forces::Dict{String, Float64}, L::Int, 
                    rng::AbstractRNG=Xoshiro(123))

Sample a sequence of length L from the model specified by the parameters in fields_forces. The sampling is exact (that is the correct probability distribution is sampled) and a random generator can be specified via rng.